Path Planning, Warnings, Comments, etc
Path Planning Completed, Demo State Created

Path Planning locates a
path from a start point to an end point.

Demo state has been created
and cleaned, needs implementing to support
game like environment.

Comments, Warnings, Walls, Collision, Positions

Final comments
added.

Warnings fixed.

Disabling of force walls setting added /
fixed

Collision detection altered to use a square area (still a bit
glitchy)

Position highlights added for start point and prepared for end
point.

Commenting, error checking

Mass commenting of everything, fixed several
errors from mouse clicking
etc, some optimisation too

+++ SFMLGame/Src/Vector2D.h
@@ -4,7 +4,7 @@
 
 class CVector2D {
 public:
-	CVector2D(int x, int y) : xPos_(x), yPos_(y) {};
+	CVector2D(float x, float y) : xPos_(x), yPos_(y) {};
 	~CVector2D();
 
 	float xPos_;
@@ -14,6 +14,7 @@ public:
 	float distance(CVector2D posB);
 	void normalize();
 
+	//Operator definitions.
 	CVector2D& operator+=(CVector2D vectorToAdd);
 	CVector2D& operator-=(CVector2D vectorToSubtract);
